Renaming the client as part of PyPI packaging effort
name: manageiq-api-client-python -> manageiq-client
package: miqclient -> manageiq_client
