+++ tox.ini
@@ -6,7 +6,7 @@ deps=
     pytest
     pytest-cov
     coveralls
-commands = py.test {posargs: -v --cov miqclient}
+commands = py.test {posargs: -v --cov manageiq_client}
 
 [testenv:codechecks]
 skip_install = true
